게시판 실행 코드입니다.

+++ src/main/webapp/BulletinBoard/delete.jsp
@@ -3,16 +3,16 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"
     import="java.sql.*"%>
-    
-<%		
-	String bcode = request.getParameter("bcode");
-	LoginDao dao = new LoginDao();
-	dao.delete(Integer.parseInt(bcode));
+
+<%
+    String bcode = request.getParameter("bcode");
+    if (bcode != null && !bcode.isEmpty()) {
+        LoginDao dao = new LoginDao();
+        dao.delete(Integer.parseInt(bcode));
+    }
 %>
 
 <script type="text/javascript">
-	let ans = alert("삭제되었습니다!");
-	if(!ans){
-		location.href ='list.jsp';
-	}
+    alert("삭제되었습니다!");
+    location.href = 'list.jsp';
 </script>
